E2E SYA update
